+++ CHANGELOG.md
@@ -6,6 +6,42 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 ### Unreleased
 n/a
 
+## 0.10.0
+
+### Change
+* Shared Components
+   * "load more" button component released with new design
+   * page snackbar component released with new design
+   * page snackbar auto closure implemented (as variant)
+   * page snackbar positioning updated
+   * tooltip enhanced by adding light color variant
+* Partner Network
+   * stop throbber in case of error and information "no rows" added
+   * empty details overlay issue fix by updating the front end business logic of the api key values fetched
+
+### Feature
+* Service Marketplace
+   * new UI released
+   * enabled different views by service category
+   * enabled sorting and in-page search
+* Connector Registration
+   * style updated and connector details enhanced inside the "my connectors" table
+   * added DAPS logic and file upload inside the connector registration ui and business logic
+* Notifications Service
+   * enabled sorting and filtering
+   * pagination enabled by adding "load more" function and receiving pagination details via api
+
+### Technical Support
+* n/a
+
+### Bugfix
+* Technical User creation: Removed multi overlay layering and corrected response style
+* Notification Service:
+   * modal for message deletion page positioning fixed
+   * sort icon behavior on click/mouse over fixed
+   * notification api trigger logic fixed
+* App user management: user role deletion fixed
+
 ## 0.9.0
 
 ### Change